refactor(domain): Generator 패키지 이동
