+++ src/main/java/domain/board/generator/IntersectionGenerator.java
@@ -1,4 +1,4 @@
-package domain.board;
+package domain.board.generator;
 
 import domain.intersection.Intersection;
