+++ src/test/java/fixture/TestIntersectionGenerator.java
@@ -1,6 +1,6 @@
 package fixture;
 
-import domain.board.IntersectionGenerator;
+import domain.board.generator.IntersectionGenerator;
 import domain.intersection.Intersection;
 
 import java.util.List;
